Combine all new test cases to one case.
