--- tests/data/preview/newline_after_def1.py
@@ -1,8 +1,0 @@
-def foo():
-
-    print("The newline above me should be deleted!")
-
-# output
-
-def foo():
-    print("The newline above me should be deleted!")
