--- tests/data/preview/newline_after_def2.py
@@ -1,10 +1,0 @@
-def foo():
-
-
-
-    print("All the newlines above me should be deleted!")
-
-# output
-
-def foo():
-    print("All the newlines above me should be deleted!")
