--- tests/data/preview/newline_after_def3.py
@@ -1,12 +1,0 @@
-def f():
-
-    print("No newline above me!")
-
-    print("There is a newline above me, and that's OK!")
-
-# output
-
-def f():
-    print("No newline above me!")
-
-    print("There is a newline above me, and that's OK!")
