--- tests/data/preview/newline_after_def4.py
@@ -1,12 +1,0 @@
-def foo():
-
-    # There is a comment here
-
-    print("The newline above me should not be deleted!")
-
-# output
-
-def foo():
-    # There is a comment here
-
-    print("The newline above me should not be deleted!")
