--- tests/data/preview/newline_after_def5.py
@@ -1,10 +1,0 @@
-class Foo:
-    def bar(self):
-
-        print("The newline above me should be deleted!")
-
-# output
-
-class Foo:
-    def bar(self):
-        print("The newline above me should be deleted!")
