--- tests/data/preview/newline_after_for1.py
@@ -1,8 +1,0 @@
-for i in range(5):
-
-    print(f"{i}) The line above me should be removed!")
-
-# output
-
-for i in range(5):
-    print(f"{i}) The line above me should be removed!")
