--- tests/data/preview/newline_after_for2.py
@@ -1,10 +1,0 @@
-for i in range(5):
-
-
-
-    print(f"{i}) The lines above me should be removed!")
-
-# output
-
-for i in range(5):
-    print(f"{i}) The lines above me should be removed!")
