--- tests/data/preview/newline_after_for3.py
@@ -1,11 +1,0 @@
-for i in range(5):
-
-    for j in range(7):
-
-        print(f"{i}) The lines above me should be removed!")
-
-# output
-
-for i in range(5):
-    for j in range(7):
-        print(f"{i}) The lines above me should be removed!")
