--- tests/data/preview/newline_after_if1.py
@@ -1,11 +1,0 @@
-import random
-
-if random.randint(0, 3) == 0:
-
-    print("The new line above me is about to be removed!")
-
-# output
-import random
-
-if random.randint(0, 3) == 0:
-    print("The new line above me is about to be removed!")
