--- tests/data/preview/newline_after_if2.py
@@ -1,14 +1,0 @@
-import random
-
-if random.randint(0, 3) == 0:
-
-
-
-
-    print("The new lines above me is about to be removed!")
-
-# output
-import random
-
-if random.randint(0, 3) == 0:
-    print("The new lines above me is about to be removed!")
