--- tests/data/preview/newline_after_if3.py
@@ -1,14 +1,0 @@
-import random
-
-if random.randint(0, 3) == 0:
-
-    if random.uniform(0, 1) > 0.5:
-
-        print("Two lines above me are about to be removed!")
-
-# output
-import random
-
-if random.randint(0, 3) == 0:
-    if random.uniform(0, 1) > 0.5:
-        print("Two lines above me are about to be removed!")
