--- tests/data/preview/newline_after_while1.py
@@ -1,8 +1,0 @@
-while True:
-
-    print("The newline above me should be deleted!")
-
-# output
-
-while True:
-    print("The newline above me should be deleted!")
