--- tests/data/preview/newline_after_while2.py
@@ -1,10 +1,0 @@
-while True:
-
-
-
-    print("The newlines above me should be deleted!")
-
-# output
-
-while True:
-    print("The newlines above me should be deleted!")
