--- tests/data/preview/newline_after_while3.py
@@ -1,11 +1,0 @@
-while True:
-
-    while False:
-
-        print("The newlines above me should be deleted!")
-
-# output
-
-while True:
-    while False:
-        print("The newlines above me should be deleted!")
