--- tests/data/preview/newline_after_with1.py
@@ -1,8 +1,0 @@
-with open("/path/to/file.txt", mode="w") as file:
-
-    file.write("The new line above me is about to be removed!")
-
-# output
-
-with open("/path/to/file.txt", mode="w") as file:
-    file.write("The new line above me is about to be removed!")
