--- tests/data/preview/newline_after_with2.py
@@ -1,10 +1,0 @@
-with open("/path/to/file.txt", mode="w") as file:
-
-
-
-    file.write("The new lines above me is about to be removed!")
-
-# output
-
-with open("/path/to/file.txt", mode="w") as file:
-    file.write("The new lines above me is about to be removed!")
